Remove proto path special-casing for schema_export
Since we use the same grpc_web rule for both hand-written protos and
protos generated from the database schema, a convention that matches
both cases for proto paths is needed. For now, we opt to include the
schema in the filename instead of a parent directory.

Later we may reintroduce a directory, but that needs to be done in a way
that is compatible with hand-written protos too.